Version mezcla
Programacion de esc mezclada en c++
